+++ gateway-android/src/main/java/com/mastercard/gateway/android/sdk/Gateway3DSecureActivity.java
@@ -13,6 +13,9 @@ import android.webkit.WebChromeClient;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
+import java.util.Map;
+import java.util.Set;
+
 
 public class Gateway3DSecureActivity extends AppCompatActivity implements Gateway3DSecureView {
 
@@ -28,22 +31,9 @@ public class Gateway3DSecureActivity extends AppCompatActivity implements Gatewa
     public static final String EXTRA_TITLE = "com.mastercard.gateway.android.TITLE";
 
     /**
-     * The 3-D Secure Id returned from the Gateway on the Process ACS Result call
-     * Will not be set on error
+     * The ACS Result data after performing 3DS
      */
-    public static final String EXTRA_3D_SECURE_ID = "com.mastercard.gateway.android.3D_SECURE_ID";
-
-    /**
-     * The summary status returned from the Gateway on the Process ACS Result call
-     * Will not be set on error
-     */
-    public static final String EXTRA_SUMMARY_STATUS = "com.mastercard.gateway.android.SUMMARY_STATUS";
-
-    /**
-     * A message indicating there was an error reading response data.
-     * Will not be set on success
-     */
-    public static final String EXTRA_ERROR = "com.mastercard.gateway.android.ERROR";
+    public static final String EXTRA_ACS_RESULT = "com.mastercard.gateway.android.ACS_RESULT";
 
 
     Toolbar toolbar;
@@ -128,28 +118,15 @@ public class Gateway3DSecureActivity extends AppCompatActivity implements Gatewa
     }
 
     @Override
-    public void error(int errorResId) {
-        Intent intent = new Intent();
-        intent.putExtra(EXTRA_ERROR, getString(errorResId));
-
-        complete(intent);
-    }
-
-    @Override
-    public void success(String summaryStatus, String threeDSecureId) {
-        Intent intent = new Intent();
-        intent.putExtra(EXTRA_SUMMARY_STATUS, summaryStatus);
-        intent.putExtra(EXTRA_3D_SECURE_ID, threeDSecureId);
-
-        complete(intent);
-    }
-
-    @Override
     public void cancel() {
         onBackPressed();
     }
 
-    void complete(Intent intent) {
+    @Override
+    public void complete(String acsResult) {
+        Intent intent = new Intent();
+        intent.putExtra(EXTRA_ACS_RESULT, acsResult);
+
         setResult(Activity.RESULT_OK, intent);
         finish();
     }